Refactore of the build system (#322)
Signed-off-by: Taras Drozdovskyi <t.drozdovsky@samsung.com>

After request from external developer on supporting the Alpine container instead of Ubuntu Vitalii55' branch, it became obvious to me that we need to configure a more flexible system for configuring the build's parameters. I suggest abandoning the management of the build using the build.sh script, and fully use Makefile. For convenience, the configuration is used standard tool Kconfig and its configuration files. This is the first step to consider my proposal by the community. I am very interested in your opinion on this issue.